Move and extend short option group tests. Fix bug uncovered by new test.
